FIX Don't try to write to script dir
Previously the script would determine
its own location instead of using the
current working directory, in contrast
to what the comment above that line
of code expressed.

Trying to create a directory at the
location of the script itself is not
a good idea anyway and will usually
fail if the script is placed e. g. in
a place like /usr/local/bin.