ENH Add backup rotation, credential file support
New options -r (rotating backups, keeping up to n
previous files) and -c for providing a mysql
configuration file with credentials to be used.
Allows storing these in a protected file instead
of having to provide it via script parameter
(which might be logged somewhere) or requiring
interactive entry.

Using the --defaults-extra-file option for the
mysql clients requires some working around
quoting issues (hence eval). Moreover, the option
needs to be the first.